# api/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Student Sync API
  description: API documentation for Student Sync application
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
...
  /user:
    get:
      tags:
        - User
      summary: Get all users
      description: Returns a list of users; supports optional pagination.
      parameters:
        - name: paginate
          in: query
          description: Enable pagination.
          schema:
            type: boolean
        - name: page
          in: query
          description: Page number.
          schema:
            type: integer
        - name: size
          in: query
          description: Page size.
          schema:
            type: integer
      responses:
        "200":
          description: List of users retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserList'
    post:
      tags:
        - User
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        "201":
          description: User created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
  /user/{id}:
    get:
      tags:
        - User
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          description: User ID.
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
    put:
      tags:
        - User
      summary: Update a user
      parameters:
        - name: id
          in: path
          description: User ID.
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        "200":
          description: User updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
    delete:
      tags:
        - User
      summary: Delete a user
      parameters:
        - name: id
          in: path
          description: User ID.
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /user/u/{username}:
    get:
      tags:
        - User
      summary: Get a user by username
      parameters:
        - name: username
          in: path
          description: Username.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
...
components:
  schemas:
    ApiResponse:
      type: object
      description: Standard API response wrapper.
      properties:
        status:
          type: string
          description: HTTP status of the response.
          example: OK
        message:
          type: string
          description: Additional message.
        timestamp:
          type: string
          format: date
          description: Response timestamp.
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'
        data:
          description: Response data.
          oneOf:
            - type: object
            - type: array
    PaginationMetadata:
      type: object
      description: Pagination metadata.
      properties:
        totalElements:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
        currentPage:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 10
...
    User:
      type: object
      description: User entity.
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserDTO:
      type: object
      description: Data transfer object for User.
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
        departmentId:
          type: string
          format: uuid
        batchId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        admissionYear:
          type: integer
        currentSemester:
          type: integer
        status:
          type: string
    ApiResponseUser:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/User'
    ApiResponseUserList:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/User'
...